Rename session service folder
