+++ app/shared/index.ts
@@ -1,6 +1,6 @@
 export * from "./bluetooth.service";
 export * from "./overview.service";
-export * from "./session-service";
+export * from "./session";
 export * from "./util";
 export * from "./login";
 export * from "./user";
