+++ app/shared/session/session.model.ts
